ch(jwt): validation - Protects all API with JWT [Delivers #163673633]

+++ model/partyQueries/deleteParty.js
@@ -1,6 +1,21 @@
 import db from '../db/db';
 
 const deleteParty = (req, res) => {
+  if (req.user.isadmin === false) {
+    return res.status(403).json({
+      status: 403,
+      error: 'Only admins are authorized to view this page.',
+    });
+  }
+
+  const isValid = /[1-9]+/.test(req.params.id);
+  if (!isValid) {
+    return res.status(400).json({
+      status: 400,
+      error: 'Request parameter must be an integer',
+    });
+  }
+
   const id = parseInt(req.params.id, 10);
 
   const text = 'DELETE FROM party WHERE id=$1 RETURNING *';
